[PORT FROM MAIN] There is a hidden issue about overflow
BZ: 77505

Rootcause: There is a hidden issue about overflow. When app want to
create it's drwaing cache. if it's width or height is large. The size
of the bitmap will be width*height*4. The size will beyond the range
of integer.So convert int to long first to avoid overflow

Oringin: Internal

Orig-Change-Id: I868652e8bc6c96fa83c752526ef189418f6653b7
Category: aosp improvement
Domain: AOSP-Framework-GUI
Origin: internal
Upstream-Candidate: no, needs rework

+++ core/java/android/view/View.java
@@ -13565,7 +13565,8 @@ public class View implements Drawable.Callback, KeyEvent.Callback,
             final boolean opaque = drawingCacheBackgroundColor != 0 || isOpaque();
             final boolean use32BitCache = attachInfo != null && attachInfo.mUse32BitDrawingCache;
 
-            final long projectedBitmapSize = width * height * (opaque && !use32BitCache ? 2 : 4);
+            final long projectedBitmapSize = (long) width * height * (opaque
+                    && !use32BitCache ? 2 : 4);
             final long drawingCacheSize =
                     ViewConfiguration.get(mContext).getScaledMaximumDrawingCacheSize();
             if (width <= 0 || height <= 0 || projectedBitmapSize > drawingCacheSize) {
